+++ lib/common.ts
@@ -126,6 +126,7 @@ function flagsForBuildOptions(options: types.BuildOptions, isTTY: boolean, logLe
   let loader = getFlag(options, keys, 'loader', mustBeObject);
   let outExtension = getFlag(options, keys, 'outExtension', mustBeObject);
   let publicPath = getFlag(options, keys, 'publicPath', mustBeString);
+  let inject = getFlag(options, keys, 'inject', mustBeArray);
   let entryPoints = getFlag(options, keys, 'entryPoints', mustBeArray);
   let stdin = getFlag(options, keys, 'stdin', mustBeObject);
   let write = getFlag(options, keys, 'write', mustBeBoolean) !== false;
@@ -143,6 +144,7 @@ function flagsForBuildOptions(options: types.BuildOptions, isTTY: boolean, logLe
   if (publicPath) flags.push(`--public-path=${publicPath}`);
   if (mainFields) flags.push(`--main-fields=${mainFields.join(',')}`);
   if (external) for (let name of external) flags.push(`--external:${name}`);
+  if (inject) for (let path of inject) flags.push(`--inject:${path}`);
   if (loader) {
     for (let ext in loader) {
       if (ext.indexOf('=') >= 0) throw new Error(`Invalid extension: ${ext}`);